add gdt

+++ kernel/include/types.h
@@ -1,14 +1,29 @@
+/*
+ * 
+ *
+ *    /--------------------------------------------O
+ *    |                                            |
+ *    |  COPYRIGHT : (c) 2024 per Linuxperoxo.     |
+ *    |  AUTHOR    : Linuxperoxo                   |
+ *    |  FILE      : types.h                       |
+ *    |  SRC MOD   : 24/11/2024                    | 
+ *    |                                            |
+ *    O--------------------------------------------/
+ *    
+ *
+ */
+
 #ifndef __TYPES__
 #define __TYPES__
 
 typedef unsigned char __u8;
 typedef unsigned short __u16;
-typedef unsigned int __u32;
-typedef unsigned long long __u64;
+typedef unsigned long int __u32;
+typedef unsigned long long int __u64;
 
 typedef signed char __i8;
 typedef signed short __i16;
-typedef signed int __i32;
-typedef signed long long __i64;
+typedef signed long int __i32;
+typedef signed long long int __i64;
 
 #endif
